knowledge_base_service/pyproject.toml
[project]
name = "knowledge-base-service"
version = "0.1.0"
description = "RAG and knowledge base management service"
requires-python = ">=3.11"
dependencies = [
    # Base dependencies
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    # Document processing
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    # Embeddings
    "sentence-transformers>=2.2.2",
    # PyTorch - required for sentence-transformers (embedding models) and reranking
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    # Reranking (optional - only needed if RERANKER_ENABLED=true)
    "transformers>=4.30.0",  # Required for reranking models (CrossEncoder) - requires torchvision
    # LLM (OpenRouter/Gemini)
    "openai>=1.3.0",  # Required for OpenRouter API communication
    # Vector databases (all supported, use VECTOR_DB_TYPE to choose)
    "qdrant-client>=1.7.0",
    "fastembed==0.7.3",  # Required for Qdrant sparse vector (BM25) generation
    "weaviate-client>=4.0.0",
    "chromadb>=0.4.18",
]

[tool.uv]
package = false

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu118" }
torchvision = { index = "pytorch-cu118" }
torchaudio = { index = "pytorch-cu118" }

[tool.hatch.build.targets.wheel]
# Include all Python packages/modules in the service
packages = ["app", "domain", "infrastructure"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

